WIP: refactor for separate stores
