+++ build.rs
@@ -3,10 +3,4 @@ fn main() {
         "cargo:rustc-env=GIT_VERSION={}",
         git_version::git_version!(args = ["--tags", "--always", "--dirty=-modified"])
     );
-
-    capnpc::CompilerCommand::new()
-        .src_prefix("schema")
-        .file("schema/schema.capnp")
-        .run()
-        .expect("schema compiler command");
 }
