+++ src/fs/mod.rs
@@ -2,13 +2,12 @@
 #![deny(clippy::unimplemented, clippy::todo)]
 
 use crate::cache;
-use crate::meta;
-use crate::meta::inode::Inode;
-use crate::meta::Entry;
-use crate::meta::Metadata;
+use crate::fungi::{
+    meta::{FileType, Inode},
+    Reader,
+};
 
 use anyhow::{ensure, Result};
-use meta::types::EntryKind;
 use polyfuse::reply::FileAttr;
 use polyfuse::{
     op,
@@ -33,13 +32,13 @@ type BlockSize = u64;
 
 #[derive(Clone)]
 pub struct Filesystem {
-    meta: meta::Metadata,
+    meta: Reader,
     cache: cache::Cache,
     lru: Arc<Mutex<lru::LruCache<FHash, (File, BlockSize)>>>,
 }
 
 impl Filesystem {
-    pub fn new(meta: meta::Metadata, cache: cache::Cache) -> Filesystem {
+    pub fn new(meta: Reader, cache: cache::Cache) -> Filesystem {
         Filesystem {
             meta,
             cache,
@@ -96,25 +95,24 @@ impl Filesystem {
     }
 
     async fn readlink(&self, req: &Request, op: op::Readlink<'_>) -> Result<()> {
-        let entry = self.meta.entry(op.ino()).await?;
-        let link = match entry.kind {
-            EntryKind::Link(l) => l,
-            _ => {
-                return Ok(req.reply_error(libc::ENOLINK)?);
-            }
-        };
+        let link = self.meta.inode(op.ino()).await?;
+        if !link.mode.is(FileType::Link) {
+            return Ok(req.reply_error(libc::ENOLINK)?);
+        }
 
-        req.reply(link.target)?;
-        Ok(())
+        if let Some(target) = link.data {
+            req.reply(target);
+            return Ok(());
+        }
+
+        return Ok(req.reply_error(libc::ENOLINK)?);
     }
 
     async fn read(&self, req: &Request, op: op::Read<'_>) -> Result<()> {
-        let entry = self.meta.entry(op.ino()).await?;
-        let file_metadata = match entry.kind {
-            EntryKind::File(file) => file,
-            _ => {
-                return Ok(req.reply_error(libc::EISDIR)?);
-            }
+        let entry = self.meta.inode(op.ino()).await?;
+
+        if !entry.mode.is(FileType::Regular) {
+            return Ok(req.reply_error(libc::EISDIR)?);
         };
 
         let offset = op.offset() as usize;
@@ -122,7 +120,9 @@ impl Filesystem {
         let chunk_size = CHUNK_SIZE; // file.block_size as usize;
         let chunk_index = offset / chunk_size;
 
-        if chunk_index >= file_metadata.blocks.len() || op.size() == 0 {
+        let blocks = self.meta.blocks(op.ino()).await?;
+
+        if chunk_index >= blocks.len() || op.size() == 0 {
             // reading after the end of the file
             let data: &[u8] = &[];
             return Ok(req.reply(data)?);
@@ -133,7 +133,7 @@ impl Filesystem {
         let mut buf: Vec<u8> = vec![0; size];
         let mut total = 0;
 
-        'blocks: for block in file_metadata.blocks.iter().skip(chunk_index) {
+        'blocks: for block in blocks.iter().skip(chunk_index) {
             // hash works as a key inside the LRU
             let hash = block.hash;
 
@@ -198,32 +198,25 @@ impl Filesystem {
     }
 
     async fn getattr(&self, req: &Request, op: op::Getattr<'_>) -> Result<()> {
-        let entry = self.meta.entry(op.ino()).await?;
+        let entry = self.meta.inode(op.ino()).await?;
+
         let mut attr = AttrOut::default();
 
         let fill = attr.attr();
-        if entry.fill(&self.meta, fill).await.is_err() {
-            req.reply_error(libc::ENOENT)?;
-        }
+        entry.fill(fill);
 
-        if op.ino() == 1 {
-            fill.mode(libc::S_IFDIR | 0o755);
-        }
         req.reply(attr)?;
 
         Ok(())
     }
 
     async fn readdir(&self, req: &Request, op: op::Readdir<'_>) -> Result<()> {
-        let entry = self.meta.entry(op.ino()).await?;
+        let root = self.meta.inode(op.ino()).await?;
 
-        let dir = match entry.kind {
-            EntryKind::Dir(dir) => dir,
-            _ => {
-                req.reply_error(libc::ENOTDIR)?;
-                return Ok(());
-            }
-        };
+        if !root.mode.is(FileType::Dir) {
+            req.reply_error(libc::ENOTDIR)?;
+            return Ok(());
+        }
 
         let mut out = ReaddirOut::new(op.size() as usize);
         let mut offset = op.offset();
@@ -233,7 +226,7 @@ impl Filesystem {
                 "..".as_ref(),
                 match op.ino() {
                     1 => 1,
-                    _ => self.meta.dir_inode(dir.parent).await?.ino(),
+                    _ => root.parent,
                 },
                 libc::DT_DIR as u32,
                 2,
@@ -244,30 +237,21 @@ impl Filesystem {
             offset -= 2;
         }
 
-        for (i, entry) in dir.entries.iter().enumerate().skip(offset as usize) {
+        let children = self.meta.children(root.ino, op.size(), offset).await?;
+        for (i, entry) in children.into_iter().enumerate() {
             let offset = i as u64 + 3;
-            let full = match &entry.kind {
-                EntryKind::SubDir(sub) => {
-                    let inode = self.meta.dir_inode(&sub.key).await?;
-                    out.entry(
-                        entry.node.name.as_ref(),
-                        inode.ino(),
-                        libc::DT_DIR as u32,
-                        offset,
-                    )
+
+            let full = match entry.mode.file_type() {
+                FileType::Dir => {
+                    //let inode = self.meta.dir_inode(&sub.key).await?;
+                    out.entry(entry.name.as_ref(), entry.ino, libc::DT_DIR as u32, offset)
                 }
-                EntryKind::File(_) => out.entry(
-                    entry.node.name.as_ref(),
-                    entry.node.inode.ino(),
-                    libc::DT_REG as u32,
-                    offset,
-                ),
-                EntryKind::Link(_) => out.entry(
-                    entry.node.name.as_ref(),
-                    entry.node.inode.ino(),
-                    libc::DT_LNK as u32,
-                    offset,
-                ),
+                FileType::Regular => {
+                    out.entry(entry.name.as_ref(), entry.ino, libc::DT_REG as u32, offset)
+                }
+                FileType::Link => {
+                    out.entry(entry.name.as_ref(), entry.ino, libc::DT_LNK as u32, offset)
+                }
                 _ => {
                     warn!("unkonwn entry");
                     false
@@ -283,28 +267,31 @@ impl Filesystem {
     }
 
     async fn lookup(&self, req: &Request, op: op::Lookup<'_>) -> Result<()> {
-        let parent = self.meta.entry(op.parent()).await?;
-
-        let dir = match parent.kind {
-            EntryKind::Dir(dir) => dir,
-            _ => {
+        let name = match op.name().to_str() {
+            Some(name) => name,
+            None => {
                 req.reply_error(libc::ENOENT)?;
                 return Ok(());
             }
         };
 
-        for entry in dir.entries.iter() {
-            if entry.node.name.as_bytes() == op.name().as_bytes() {
-                let mut out = EntryOut::default();
-                let inode = entry.fill(&self.meta, out.attr()).await?;
-                out.ino(inode.ino());
-                out.ttl_attr(TTL);
-                out.ttl_entry(TTL);
-                return Ok(req.reply(out)?);
-            }
-        }
+        let node = self.meta.lookup(op.parent(), name).await?;
 
-        Ok(req.reply_error(libc::ENOENT)?)
+        let node = match node {
+            Some(node) => node,
+            None => {
+                req.reply_error(libc::ENOENT)?;
+                return Ok(());
+            }
+        };
+        let mut out = EntryOut::default();
+
+        node.fill(out.attr());
+        out.ino(node.ino);
+        out.ttl_attr(TTL);
+        out.ttl_entry(TTL);
+
+        return Ok(req.reply(out)?);
     }
 }
 
@@ -346,60 +333,25 @@ impl AsyncSession {
     }
 }
 
-#[async_trait::async_trait]
 trait AttributeFiller {
-    async fn fill(&self, meta: &Metadata, attr: &mut FileAttr) -> Result<Inode>;
+    fn fill(&self, attr: &mut FileAttr);
 }
 
-#[async_trait::async_trait]
-impl AttributeFiller for Entry {
-    async fn fill(&self, meta: &Metadata, attr: &mut FileAttr) -> Result<Inode> {
+impl AttributeFiller for Inode {
+    fn fill(&self, attr: &mut FileAttr) {
         use std::time::Duration;
 
-        let mode = match meta.aci(&self.node.acl).await {
-            Ok(aci) => {
-                attr.uid(aci.user);
-                attr.gid(aci.group);
-                aci.mode & 0o777
-            }
-            Err(_) => 0o444,
+        attr.mode(self.mode.mode());
+
+        attr.ino(self.ino);
+        attr.ctime(Duration::from_secs(self.ctime as u64));
+        attr.mtime(Duration::from_secs(self.mtime as u64));
+        attr.size(self.size);
+
+        match self.mode.file_type() {
+            FileType::Dir => attr.nlink(2),
+            FileType::Regular => attr.blksize(4 * 1024),
+            _ => (),
         };
-
-        let inode = self.node.inode;
-        attr.ino(inode.ino());
-        attr.ctime(Duration::from_secs(self.node.creation as u64));
-        attr.mtime(Duration::from_secs(self.node.modification as u64));
-        attr.size(self.node.size);
-
-        let inode = match &self.kind {
-            EntryKind::Unknown => bail!("unkown entry"),
-            EntryKind::Dir(_) => {
-                attr.nlink(2);
-                attr.mode(libc::S_IFDIR | mode);
-                inode
-            }
-            EntryKind::SubDir(sub) => {
-                let inode = meta.dir_inode(&sub.key).await?;
-                // reset inode
-                attr.ino(inode.ino());
-                attr.nlink(2);
-                attr.mode(libc::S_IFDIR | mode);
-                inode
-            }
-            EntryKind::File(_) => {
-                attr.nlink(1);
-                attr.mode(libc::S_IFREG | mode);
-                attr.blksize(4 * 1024);
-                inode
-            }
-            EntryKind::Link(link) => {
-                attr.nlink(1);
-                attr.size(link.target.len() as u64);
-                attr.mode(libc::S_IFLNK | 0o555);
-                inode
-            }
-        };
-
-        Ok(inode)
     }
 }
